Changed e-fuse code data type

+++ HWDescription/RD53.cc
@@ -70,7 +70,7 @@ RD53::RD53(uint8_t            pBeId,
            uint8_t            pHybridId,
            uint8_t            pRD53Id,
            uint8_t            pRD53Lane,
-           uint8_t            pRD53eFuseCode,
+           uint32_t           pRD53eFuseCode,
            const std::string& fileName,
            const std::string& cfgComment)
     : ReadoutChip(pBeId, pFMCId, pOpticalGroupId, pHybridId, pRD53Id)
