+++ HWDescription/RD53B.cc
@@ -146,7 +146,7 @@ RD53B::RD53B(const FrontEndType& frontEndType,
              uint8_t             pHybridId,
              uint8_t             pRD53Id,
              uint8_t             pRD53Lane,
-             uint8_t             pRD53eFuseCode,
+             uint32_t            pRD53eFuseCode,
              const std::string&  fileName,
              const std::string&  cfgComment)
     : RD53(pBeId, pFMCId, pOpticalGroupId, pHybridId, pRD53Id, pRD53Lane, pRD53eFuseCode, fileName, cfgComment)
